+++ PartyGame/Views/PhraseView.swift
@@ -9,78 +9,107 @@ import SwiftUI
 
 struct PhraseView: View {
     
-    @State var selectedPhrase: Phrase = .init(text: "", category: .action)
+    @State var selectedPhrase: Phrase = .init(text: "write your own phrase here", category: .action)
     @State var displayedPhrases: [Phrase] = []
-    var viewModel = PhraseViewModel()
-    
-    @State var nextScreen: Bool = false
     
     let columns = [GridItem(.flexible())]
+    
     var body: some View {
-        NavigationStack {
-            VStack (spacing: 400) {
+        ZStack{
+            Image("texture")
+                .resizable()
+                .scaledToFill()
+                .frame(minWidth: 0)
+                .edgesIgnoringSafeArea(.all)
+                
+            VStack (spacing: 101) {
+                Text("Component")
+                    .frame(alignment: .top)
+                    .font(.system(size: 64, weight: .bold, design: .default))
                 VStack(spacing: 64){
-                    ZStack(alignment: .trailing){
+                    VStack(spacing: 16){
                         TextField(
-                            "Write down a phrase or select one",
+                            "write your own phrase here",
                             text: $selectedPhrase.text)
-                        
+                        .foregroundStyle(.lilac)
                         .padding(.vertical, 12)
                         .padding(.horizontal, 16)
-                        .frame(maxWidth: .infinity)
+                        .frame(maxWidth: .infinity, minHeight: 52, maxHeight: 52, alignment: .leading)
+                        .background(RoundedRectangle(cornerRadius: 26)
+                            .fill(Color.lighterPurple.shadow(.inner(color: .darkerPurple, radius: 2, y: 3))))
+                    
+                        VStack(spacing: 16){
+                            HStack(spacing: 63){
+                                Text("or choose one of ours:")
+                                    .foregroundStyle(.ice)
+                                    .fontWeight(.semibold)
+                                    .font(.system(size: 20, weight: .semibold, design: .rounded))
+                                Button{
+                                    displayedPhrases = Array(Phrases.all.shuffled().prefix(3))
+                                }label: {
+                                    Image("Dice")
+                                        .background(Circle()
+                                            .frame(width: 35, height: 36)
+                                            .foregroundStyle(Color.lilac))
+                                            
+                                }
+                                       
+                            }
+                            LazyVGrid(columns: columns, spacing: 8){
+                                ForEach(displayedPhrases, id: \.self) { phrase in
+                                    PhraseComponent(phrase: phrase,
+                                                    isSelected: phrase == selectedPhrase,
+                                                    onSelect: {
+                                        selectedPhrase = phrase
+                                    })
+                                }
+                                
+                            }
+                        }
+                        .padding(.vertical, 12)
+                        .padding(.horizontal)
+                        .background(RoundedRectangle(cornerRadius: 26).fill(Color.lighterPurple))
                         
-                        .overlay(
-                            RoundedRectangle(cornerRadius: 16)
-                                .stroke(.black, lineWidth: 0.5)
-                        )
-                        Button{
-                            print("Selected random phrase")
-                            displayedPhrases = Array(Phrases.all.shuffled().prefix(4))
-                        } label:{
-                            Image(systemName: "dice.fill")
-                                .foregroundStyle(.black)
-                                .font(.system(size: 24))
-                                .frame(width: 50, height: 50)
-                        }
                     }
-                    LazyVGrid(columns: columns, spacing: 16){
-                        ForEach(displayedPhrases, id: \.self) { phrase in
-                            PhraseComponent(phrase: phrase,
-                                            isSelected: phrase == selectedPhrase,
-                                            onSelect: {
-                                selectedPhrase = phrase
-                            })
-                        }
-                        
-                    }
-                    Button {
-                        viewModel.submitPhrase(phrase: selectedPhrase.text)
-                        if viewModel.haveAllPlayersSubmitted {
-                            nextScreen = true
-                        }
-                    }label: {
-                        Text("Submit Phrase")
-                    }
-                    .foregroundStyle(.primary)
-                    .background(
-                        RoundedRectangle(cornerRadius: 16)
-                            .fill(.tertiary)
-                            .frame(width: 356, height: 42)
-                    )
+                    .padding(.horizontal)
+                    .padding(.vertical)
+                    .background(GradientBackground()
+                    /*.opacity(0.5)*/)
                     
                 }
+                .padding(.horizontal)
+                Button {
+                    print("Submitted phrase: \(selectedPhrase)")
+                }label: {
+                    Text("Component")
+                }
+                .foregroundStyle(.primary)
+                .background(
+                    RoundedRectangle(cornerRadius: 16)
+                        .fill(.tertiary)
+                        .frame(width: 356, height: 42)
+                )
                 .padding(.horizontal, 16)
                 .padding(.vertical, 16)
                 .onAppear() {
-                    displayedPhrases = Array(Phrases.all.shuffled().prefix(4))
+                    displayedPhrases = Array(Phrases.all.shuffled().prefix(3))
                 }
             }
         }
-        .navigationBarBackButtonHidden(true)
-        .navigationDestination(isPresented: $nextScreen) {
-            ImageSelectionView()
-        }
-        
+        .background(Color.darkerPurple)
+    }
+}
+
+struct GradientBackground: View {
+    let gradientBackground = LinearGradient(
+        gradient: Gradient(colors: [.lilac.opacity(0.5), .lighterPink.opacity(0.5)]),
+        startPoint: .top,
+        endPoint: .bottom)
+
+    var body: some View {
+        RoundedRectangle(cornerRadius: 32)
+            .fill(gradientBackground
+                .shadow(.inner(color: Color.lilac, radius: 2, x: 0, y: 5)))
     }
 }
 #Preview {
